+++ test/supported-props-test.js
@@ -14,6 +14,7 @@ describe('Tokenizer Supported Properties', function () {
   , buffer: isString
   , writable: isBoolean
   , readable: isBoolean
+  , currentRule: function (v) { return v === null || isString(v) }
   }
 
   function isNumber () {